add colors to console report

+++ src/benchmark.rs
@@ -1,11 +1,11 @@
 pub trait Average {
-    fn avg(&self) -> u128;
+    fn avg(&self) -> u64;
 }
 
 impl Average for Vec<Result> {
-    fn avg(&self) -> u128 {
-        let total: u128 = self.iter().map(|r| r.duration).sum();
-        let size: u128 = self.iter().len() as u128;
+    fn avg(&self) -> u64 {
+        let total: u64 = self.iter().map(|r| r.duration).sum();
+        let size: u64 = self.iter().len() as u64;
         total / size
     }
 }
@@ -13,7 +13,7 @@ impl Average for Vec<Result> {
 #[derive(Debug)]
 pub struct Result {
     pub(crate) status: u16,
-    pub(crate) duration: u128,
+    pub(crate) duration: u64,
 }
 
 #[derive(Debug)]
@@ -29,7 +29,7 @@ impl Report {
         self.results.len()
     }
 
-    pub fn avg(&self) -> u128 {
+    pub fn avg(&self) -> u64 {
         self.results.avg()
     }
